Adding script to compute population density
This script streams in a GeoJSON file and applies a function to
each feature in the json object.  Currently it computes the area
of the polygon and computes a population density for each polygon.
The script generates a PostgreSQL script to create a new table
and insert data for each feature.

Usage: node index.js | psql <db_name>